% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExp-add.R
\name{GINIindex1D}
\alias{GINIindex1D}
\title{GINI PPindex}
\usage{
GINIindex1D(origclass,origdata,proj)
}
\arguments{
\item{origclass}{class information vector}

\item{origdata}{data matrix without class information}

\item{proj}{projection matrix}
}
\description{
GINI projection pursuit index
}
\details{
Calculate GINI projection pursuit index value
}
\examples{
data(iris)
GINIindex1D(iris[,5],as.matrix(iris[,1,drop=FALSE]))

}
\keyword{projection}
\keyword{pursuit}
