% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LDAopt.R
\name{LDA.opt.1D}
\alias{LDA.opt.1D}
\title{Find the 1D-optimal projection using LDA index for classification}
\usage{
LDA.opt.1D(i.class, i.data, weight = TRUE, ...)
}
\arguments{
\item{i.class}{class information}

\item{i.data}{data without class information}

\item{weight}{weight flag using in LDA index}
}
\value{
index.best optimal value of LDA index value

proj.best optimal 1D projection
}
\description{
Find the 1D-optimal projection using LDA index for classification
}
\examples{
data(iris)
LDA.proj.result <- LDA.opt.1D(iris[,5],iris[,1:4])
LDA.proj.result
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005)
Projection Pursuit for exploratory supervised classification,
Journal of Computational and Graphical statistics, 14(4):831-846.
}
\keyword{projection}
\keyword{pursuit}

