% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LDATree.R
\name{LDA.Tree}
\alias{LDA.Tree}
\title{Tree with LDA.}
\usage{
LDA.Tree(origclass, origdata, weight = TRUE, ...)
}
\arguments{
\item{origclass}{class information}

\item{origdata}{data  without class information}

\item{weight}{weight flag in LDA index}

\item{...}{arguments to be passed to methods}
}
\value{
Tree.Struct Tree structure of PPtree result

projbest.node 1-dim optimal projections of each split node

splitCutoff.node cutoff values of each split node

origclass original class

origdata original data
}
\description{
Find tree structure using linear discriminant(LD) in each split.
}
\examples{
data(iris)
Tree.result <- LDA.Tree(iris[,5],iris[,1:4])
Tree.result
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

