% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LDATree.R
\name{LDA.Tree.viz}
\alias{LDA.Tree.viz}
\title{Find tree structure using linear discriminant(LD) in each split.}
\usage{
LDA.Tree.viz(i.class, i.data, weight = TRUE, ...)
}
\arguments{
\item{i.class}{class information}

\item{i.data}{A training data  without class information}

\item{weight}{weight flag using in LDA index}
}
\value{
Tree.Struct Tree structure of PPtree result

Alpha.Keep 1D projections of each split

C.Keep spliting rules for each split

orig.class i.class

orig.data i.data
}
\description{
Find tree structure using linear discriminant(LD) in each split.
}
\examples{
data(iris)
Tree.result <- LDA.Tree.viz(iris[,5],iris[,1:4])
Tree.result
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

