% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExp-add.R
\name{LDAindex}
\alias{LDAindex}
\title{LDA PPindex}
\usage{
LDAindex(origclass,origdata,proj,weight=TRUE)
}
\arguments{
\item{origclass}{class information vector}

\item{origdata}{data matrix without class information}

\item{proj}{projection matrix}

\item{weight}{weight flag in LDA}
}
\description{
LDA projection pursuit index
}
\details{
Calculate LDA projection pursuit index value
}
\examples{
data(iris)
LDAindex(iris[,5],as.matrix(iris[,1:4]))
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005) 
Projection Pursuit for exploratory supervised classification, 
Journal of Computational and Graphical statistics, 14(4):831-846.
}
\keyword{projection}
\keyword{pursuit}
