% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDAopt.R
\name{PDA.opt.1D}
\alias{PDA.opt.1D}
\title{Find the 1D-optimal projection using PDA index for classification}
\usage{
PDA.opt.1D(i.class, i.data, weight=TRUE,lambda, ...)
}
\arguments{
\item{i.class}{class information}

\item{i.data}{A training data  without class information}

\item{lambda}{lambda in PDA index}

\item{weight}{weight flag using in PDA index}
}
\value{
index.best optimal value of LDA index value

proj.best optimal 1D projection
}
\description{
Find the 1D-optimal projection using PDA index for classification
}
\examples{
data(iris)
PDA.proj.result <- PDA.opt.1D(iris[,5],iris[,1:4],lambda=0.1)
PDA.proj.result
}
\references{
Lee, EK., Cook, D.(2010)
A projection pursuit index for large p small n data,
Statistics and Computing, 20:381-392.
}
\keyword{projection}
\keyword{pursuit}

