% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDATree.R
\name{PDA.Tree}
\alias{PDA.Tree}
\title{Tree with PDA}
\usage{
PDA.Tree(origclass, origdata, weight = TRUE, lambda=0.1, ...)
}
\arguments{
\item{origclass}{class information}

\item{origdata}{A training data  without class information}

\item{weight}{weight flag in PDA index}

\item{lambda}{a parameter for PDA index}

\item{...}{arguments to be passed to methods}
}
\value{
Tree.Struct Tree structure of PPtree result

projbest.node 1-dim optimal projections of each split node

splitCutoff.node cutoff values of each split node

origclass original class

origdata original data
}
\description{
Find tree structure using penalized discriminant(PD) in each split.
}
\examples{
data(iris)
Tree.result <- PDA.Tree(iris[,5],iris[,1:4],lambda=0.2)
Tree.result
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

