% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PDATree.R
\name{PDA.Tree.viz}
\alias{PDA.Tree.viz}
\title{Find tree structure using penalized discriminant(PD) in each split.}
\usage{
PDA.Tree.viz(i.class, i.data, weight = TRUE, lambda=1, ...)
}
\arguments{
\item{i.class}{class information}

\item{i.data}{A training data  without class information}

\item{weight}{weight flag using in LDA index}

\item{lambda}{a parameter for PDA index}
}
\value{
Tree.Struct Tree structure

Alpha.Keep 1D projection of each split

C.Keep spliting rule for each split

orig.class i.class

orig.data i.data
}
\description{
Find tree structure using penalized discriminant(PD) in each split.
}
\examples{
data(iris)
Tree.result <- PDA.Tree.viz(iris[,5],iris[,1:4],lambda=0.2)
Tree.result
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

