% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExp-add.R
\name{PDAindex}
\alias{PDAindex}
\title{PDA PPindex}
\usage{
PDAindex(origclass,origdata,proj,weight=TRUE,lambda=0.1)
}
\arguments{
\item{origclass}{class information vector}

\item{origdata}{data matrix without class information}

\item{proj}{projection matrix}

\item{weight}{weight flag in PDA}

\item{lambda}{lambda in PDA index}
}
\description{
PDA projection pursuit index
}
\details{
Calculate PDA projection pursuit index value
}
\examples{
data(iris)
PDAindex(iris[,5],as.matrix(iris[,1:4]),lambda=0.2)


}
\references{
Lee, EK., Cook, D.(2010) 
A projection pursuit index for large p small n data, 
Statistics and Computing, 20:381-392.
}
\keyword{projection}
\keyword{pursuit}
