% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPPtree.R
\name{plot.PPtree}
\alias{plot}
\alias{plot.PPtree}
\title{Plot PPtree}
\usage{
\method{plot}{PPtree}(x,font.size=17,width.size=1,...)
}
\arguments{
\item{x}{PPtree object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{...}{arguments to be passed to methods}
}
\description{
Plot PPtree
}
\examples{
data(iris)
Tree.result <- LDA.Tree(iris[,5],iris[,1:4])
Tree.result
plot(Tree.result)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386..
}
\keyword{tree}

