% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPPtreeclass.R
\name{plot.PPtreeclass}
\alias{plot.PPtreeclass}
\alias{plot}
\title{PPtree plot}
\usage{
\method{plot}{PPtreeclass}(
  x,
  font.size = 17,
  width.size = 1,
  main = "Projection Pursuit Classification Tree",
  sub = NULL,
  ...
)
}
\arguments{
\item{x}{PPtreeclass object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{main}{main title}

\item{sub}{sub title}

\item{...}{arguments to be passed to methods}
}
\description{
projection pursuit classification tree plot
}
\details{
Draw projection pursuit classification tree with tree structure. It is 
modified from a function in party library.
}
\examples{
data(iris)
Tree.result <- PPTreeclass(Species~., data = iris,"LDA")
Tree.result
plot(Tree.result)
}
\references{
Lee, EK(2017) 
PPtreeViz: An R Package for Visualizing Projection Pursuit Classification 
Trees, Journal of Statistical Software <doi:10.18637/jss.v083.i08>
}
\keyword{tree}
