% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPPtreereg.R
\name{plot.PPtreereg}
\alias{plot.PPtreereg}
\title{Plot PP regression tree}
\usage{
\method{plot}{PPtreereg}(x,font.size=17,width.size=1,...)
}
\arguments{
\item{x}{regPPtree object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{...}{arguments to be passed to methods}
}
\description{
plot projection pursuit regression tree structure
}
\examples{
data(mtcar)
Tree.result <- PP.Tree.reg(mtcars[,1],mtcars[,-1],final.rule=1,DEPTH=2)
plot(Tree.result)
}
\keyword{tree}

