% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPclassify.R
\name{PP.classify.viz}
\alias{PP.classify.viz}
\title{Predict class for the test set and calculate prediction error}
\usage{
PP.classify(test.data, true.class, Tree.result, Rule, ...)
}
\arguments{
\item{test.data}{the test dataset}

\item{true.class}{true class of test dataset if available}

\item{Tree.result}{the result of PP.Tree}

\item{Rule}{split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)}
}
\value{
predict.class predicted class

predict.error prediction error
}
\description{
After finding tree structure, predict class for the test set and calculate prediction error.
}
\examples{
data(iris)
n <- nrow(iris)
tot <- c(1:n)
n.train <- round(n*0.9)
train <- sample(tot,n.train)
test <- tot[-train]
Tree.result <- LDA.Tree.viz(iris[train,5],iris[train,1:4])
PP.classify.viz(iris[test,1:4],iris[test,5],Tree.result,1)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

