% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPTree.R
\name{PP.Tree}
\alias{PP.Tree}
\title{Find tree structure using various projection pursuit indices of classification in each split.}
\usage{
PP.Tree(origclass, origdata, PPmethod="LDA",weight=TRUE,r=1,lambda=0.1,
              energy=0,cooling=0.999,TOL=0.0001,maxiter=50000, ...)
}
\arguments{
\item{origclass}{class information}

\item{origdata}{A training data  without class information}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA", "Lr", "GINI", and "ENTROPY"}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{r}{r in Lr index}

\item{lambda}{lambda in PDA index}

\item{energy}{energy parameter}

\item{cooling}{cooling parameter}

\item{TOL}{tolerance}

\item{maxiter}{number of maximum iteration}

\item{...}{arguments to be passed to methods}
}
\value{
Tree.Struct Tree structure of PPtree result

projbest.node 1-dim optimal projections of each split node

splitCutoff.node cutoff values of each split node

origclass original class

origdata original data
}
\description{
Find tree structure using various projection pursuit indices of classification in each split.
}
\examples{
data(iris)
Tree.result <- PP.Tree(iris[,5],iris[,1:4],"LDA")
Tree.result
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

