% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPtreereg.R
\name{PP.Tree.reg}
\alias{PP.Tree.reg}
\title{Projection pursuit regression tree with various PP indices}
\usage{
PP.Tree.reg(origY,origX,final.rule=1,DEPTH=NULL,Rr=1,PPmethod="LDA",weight=TRUE,lambda=0.1,r=1,TOL.CV=0.1,selP=NULL,energy=0,maxiter=500,...)
}
\arguments{
\item{origY}{dependent data vector}

\item{origX}{independent data matrix without dependent variable}

\item{final.rule}{rule to calculate the final node value}

\item{DEPTH}{depth of the projection pursuit regression tree}

\item{Rr}{cutoff rule in each node}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA", "Lr", "GINI", and "ENTROPY"}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{lambda}{lambda in PDA index}

\item{r}{r in Lr index}

\item{TOL.CV}{CV limit for the final node}

\item{selP}{number of variables for the final node in Method 5}

\item{energy}{energy parameter}

\item{maxiter}{number of maximum iteration}

\item{...}{arguments to be passed to methods}
}
\value{
Tree.result projection pursuit regression tree result with PPtreeclass object format

MSE mean squared error of the final tree

mean.G  means of the observations in the final node

sd.G standard deviations of the observations in the final node.

coef.G regression coefficients for Method 3, 4 and 5

origY original dependent variable vector
}
\description{
Find tree structure using linear discriminant(LD) in each split.
}
\examples{
data(mtcars)
Tree.result <- PP.Tree.reg(mtcars[,1],mtcars[,-1],final.rule=1,DEPTH=2,PPmethod="LDA")
Tree.result
}
\references{
...
}
\keyword{tree}

