% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPTreeNew.R
\name{PP.Tree.viz}
\alias{PP.Tree.viz}
\title{Find tree structure using various projection pursuit indices of classification in each split.}
\usage{
PP.Tree.viz(PPmethod,  i.class, i.data, weight = TRUE, ...)
}
\arguments{
\item{PPmethod}{method for projection pursuit; LDA1,LDA2,PDA1,PDA2,Lp1,Lp2,GINI, ENT}

\item{i.class}{class information}

\item{i.data}{A training data  without class information}

\item{weight}{weight flag using in LDA index}
}
\value{
Tree.Struct Tree structure of PPtree result

Alpha.Keep 1D projections of each split

C.Keep spliting rules for each split

orig.class i.class

orig.data i.data
}
\description{
Find tree structure using various projection pursuit indices of classification in each split.
}
\examples{
data(iris)
Tree.result <- PP.Tree.viz("LDA1",iris[,5],iris[,1:4])
Tree.result
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

