% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPoptViz.R
\name{PPopt.Viz}
\title{Visualize PPopt}
\usage{
PPoptim.Viz(PPoptOBJ)
}
\arguments{
\item{PPoptOBJ}{result from LDAopt, PDAopt, and PPopt}
}
\description{
Visualize Projection pursuit optimization result result
}
\examples{
data(iris)
PPopt.Viz(LDAopt(iris[,5],iris[,1:4],q=1))
PPopt.Viz(LDAopt(iris[,5],iris[,1:4],q=2))
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005)
Projection Pursuit for exploratory supervised classification,
Journal of Computational and Graphical statistics, 14(4):831-846.
}
\keyword{tree}

