% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPregpredict.R
\name{PPreg.predict}
\alias{PPreg.predict}
\title{predict PPtree}
\usage{
PPreg.predict(PPtreeregOBJ,test.dataX,Rule=1,trueY=NULL,...)
}
\arguments{
\item{PPtreeregOBJ}{PPtreereg object}

\item{test.dataX}{the test dataset}

\item{Rule}{split rule 1: mean of two group means
2: weighted mean of two group means - weight with group size
3: weighted mean of two group means - weight with group sd
4: weighted mean of two group means - weight with group se
5: mean of two group medians
6: weighted mean of two group medians - weight with group size
7: weighted mean of two group median - weight with group IQR
8: weighted mean of two group median - weight with group IQR and size
9: cutoff that minimize error rates in each node}

\item{trueY}{true Y values of test dataset if available}

\item{...}{arguments to be passed to methods}
}
\value{
predict.Y predicted values

predict.MSE MSE of the predicted values
}
\description{
Predict class for the test set and calculate prediction error
After finding tree structure, predict class for the test set and calculate prediction error.
}
\examples{
data(mtcars)
n <- nrow(mtcars)
tot <- c(1:n)
n.train <- round(n*0.8)
train <- sample(tot,n.train)
test <- tot[-train]
Tree.result <- PP.Tree.reg(mtcars[train,1],mtcars[train,-1],
                           final.rule=1,DEPTH=2)
PPreg.predict(Tree.result,mtcars[test,-1],1,mtcars[test,1])
}
\keyword{tree}

