% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPregNodeViz.R
\name{PPregNode.Viz}
\alias{PPregNode.Viz}
\title{Visualization of each node in PPtree}
\usage{
PPregNode.Viz(PPtreeregOBJ,node.id,Rule)
}
\arguments{
\item{PPtreeregOBJ}{PPregtree object}

\item{node.id}{node ID}

\item{Rule}{cutoff rule}
}
\description{
Explore PPtree with different Rules in each split.
}
\examples{
data(mtcars)
Tree.result <- PP.Tree.reg(mtcars[,1],mtcars[,-1],final.rule=1,DEPTH=2)
plot(Tree.result)
PPregNode.Viz(Tree.result,1,1)
}
\keyword{tree}

