% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPtreeVizPlot.R
\name{PPtreeViz.plot}
\alias{PPtreeViz.plot}
\title{Finding PP tree structure using LDA index}
\usage{
PPtreeViz.plot(PPtreeOBJ,node.id,Rule)
}
\arguments{
\item{PPtreeOBJ}{pptree object}

\item{Rule}{cutoff rule}

\item{node.id}{node ID}
}
\description{
Find tree structure using linear discriminant in each split.
}
\examples{
data(iris)
Tree.result <- LDA.Tree.viz(iris[,5],iris[,1:4])
Tree.result
PPtreeViz.plot(Tree.result,1,1)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\seealso{
{\code{\link{PPindex.class}}, \code{\link{PP.optimize}}}
}
\keyword{tree}

