% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printPPtree.R
\name{print.PPtree}
\alias{print}
\alias{print.PPtree}
\title{Print PPtree result}
\usage{
\method{print}{PPtree}(x,coef.print=FALSE,cutoff.print=FALSE,verbose=TRUE,...)
}
\arguments{
\item{x}{PPtree object}

\item{coef.print}{print projection coefficient in each node when TRUE}

\item{cutoff.print}{print cutoff values in each node when TRUE}

\item{verbose}{print if TRUE, no output if FALSE}

\item{...}{arguments to be passed to methods}
}
\description{
Print PPtree result
}
\examples{
data(iris)
Tree.result <- LDA.Tree(iris[,5],iris[,1:4])
Tree.result
print(Tree.result,coef.print=TRUE,cutoff.print=TRUE)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

