% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printPPtreereg.R
\name{print.PPtreereg}
\alias{print.PPtreereg}
\title{Print PP.Tree.reg result}
\usage{
\method{print}{PPtreereg}(x,tree.print=TRUE,coef.print=FALSE,cutoff.print=FALSE,verbose=TRUE,...)
}
\arguments{
\item{x}{regPPtree object}

\item{tree.print}{print the tree structure when TRUE}

\item{coef.print}{print the projection coefficient in each node when TRUE}

\item{cutoff.print}{print the cutoff values in each node when TRUE}

\item{verbose}{print if TRUE, no output if FALSE}

\item{...}{arguments to be passed to methods}
}
\description{
Print the projection pursuit regression tree result
}
\examples{
data(mtcars)
Tree.result <- PP.Tree.reg(mtcars[,1],mtcars[,-1],final.rule=1,DEPTH=2,PPmethod="LDA")
Tree.result
print(Tree.result,coef.print=TRUE,cutoff.print=TRUE)
}
\keyword{tree}

