% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ENTROPYindex1D.R
\name{ENTROPYindex1D}
\alias{ENTROPYindex1D}
\title{ENTROPY PP index}
\usage{
ENTROPYindex1D(projclass, projdata)
}
\arguments{
\item{projclass}{class information}

\item{projdata}{projected data}
}
\description{
Calculate ENTROPY projection pursuit index value
}
\examples{
data(iris)
ENTROPYindex1D(iris[,5],iris[,1])
}
\keyword{projection}
\keyword{pursuit}

