% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ENTROPYindex1D}
\alias{ENTROPYindex1D}
\title{ENTROPY PPindex}
\usage{
ENTROPYindex1D(origclass,origdata,proj)
}
\arguments{
\item{origclass}{class information vector}

\item{origdata}{data matrix without class information}

\item{proj}{projection matrix}
}
\description{
ENTROPY projection pursuit index
}
\details{
Calculate ENTROPY projection pursuit index value
}
\examples{
data(iris)
ENTROPYindex1D(iris[,5],as.matrix(iris[,1,drop=FALSE]))
}
\keyword{projection}
\keyword{pursuit}

