% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GINIindex.R
\name{GINIindex1D}
\alias{GINIindex1D}
\title{GINI PP index}
\usage{
GINIindex1D(projclass,projdata)
}
\arguments{
\item{projclass}{class information}

\item{projdata}{projected data}
}
\description{
Calculate GINI projection pursuit index value
}
\examples{
data(iris)
GINIindex1D(iris[,5],iris[,1])
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005)
Projection Pursuit for exploratory supervised classification,
Journal of Computational and Graphical statistics, 14(4):831-846.
}
\keyword{projection}
\keyword{pursuit}

