% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LDAindex.R
\name{LDAindex}
\alias{LDAindex}
\title{LDA PP index}
\usage{
LDAindex(projclass,projdata,weight=TRUE)
}
\arguments{
\item{projclass}{class information}

\item{projdata}{projected data}

\item{weight}{weight flag in LDA}
}
\description{
Calculate LDA projection pursuit index value
}
\examples{
data(iris)
LDAindex(iris[,5],as.matrix(iris[,1:4]))
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005)
Projection Pursuit for exploratory supervised classification,
Journal of Computational and Graphical statistics, 14(4):831-846.
}
\keyword{projection}
\keyword{pursuit}

