% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LDAopt.R
\name{LDAopt}
\alias{LDAopt}
\title{PP optimization using LDA index}
\usage{
LDAopt(origclass, origdata,  q=1, weight = TRUE,...)
}
\arguments{
\item{origclass}{class information}

\item{origdata}{data without class information}

\item{q}{dimension of projection matrix}

\item{weight}{weight flag using in LDA index}

\item{...}{arguments to be passed to methods}
}
\value{
indexbest optimal value of LDA index value

projbest optimal q-dim projection

origclass class information

origdata A training data  without class information
}
\description{
Find the q-dim optimal projection using LDA projectin pursuit index
}
\examples{
data(iris)
LDA.proj.result <- LDAopt(iris[,5],iris[,1:4])
LDA.proj.result
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005)
Projection Pursuit for exploratory supervised classification,
Journal of Computational and Graphical statistics, 14(4):831-846.
}
\keyword{projection}
\keyword{pursuit}

