% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lrindex.R
\name{Lrindex}
\alias{Lrindex}
\title{Calculate Lr projection pursuit index value}
\usage{
Lrindex(projclass,projdata,weight=TRUE,r=1)
}
\arguments{
\item{projclass}{class information}

\item{projdata}{A training data  without class information}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{r}{r in Lr index}
}
\description{
Calculate Lr projection pursuit index value
}
\examples{
data(iris)
Lrindex(iris[,5],as.matrix(iris[,1:4]))
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005)
Projection Pursuit for exploratory supervised classification,
Journal of Computational and Graphical statistics, 14(4):831-846.
}
\keyword{projection}
\keyword{pursuit}

