% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPPtreeclass.R
\name{plot.PPtreeclass}
\alias{plot}
\alias{plot.PPtreeclass}
\title{PPtree plot}
\usage{
\method{plot}{PPtreeclass}(x, font.size = 17, width.size = 1, ...)
}
\arguments{
\item{x}{PPtreeclass object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{...}{arguments to be passed to methods}
}
\description{
projection pursuit classification tree plot
}
\details{
Draw projection pursuit classification tree with tree structure. It is
modified from a function in party library.
}
\examples{
data(iris)
Tree.result <- PP.Tree.class(iris[,5],iris[,1:4],"LDA")
Tree.result
plot(Tree.result)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

