% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPclassNodeViz.R
\name{PPclassNode.Viz}
\alias{PPclassNode.Viz}
\title{Visualization of each node in PPtree}
\usage{
PPclassNode.Viz(PPclassOBJ,node.id,Rule,legend,std,image)
}
\arguments{
\item{PPclassOBJ}{PPregclass object}

\item{node.id}{node ID}

\item{Rule}{cutoff rule}

\item{legend}{flag to represent legend in the plot. Default value is TRUE}

\item{std}{flag to standardize data before drawing plot}

\item{image}{flag to draw image plot of correlation matrix}
}
\description{
Explore PPtree with different Rules in each split.
}
\examples{
data(iris)
Tree.result <- PP.Tree.class(iris[,5],iris[,1:4],"LDA")
Tree.result
PPclassNode.Viz(Tree.result,1,1)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

