% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPoptViz.R
\name{PPoptViz}
\alias{PPoptViz}
\title{PPopt visualization}
\usage{
PPoptViz(PPoptOBJ)
}
\arguments{
\item{PPoptOBJ}{PPoptim object. result from LDAopt, PDAopt, and PPopt}
}
\description{
Visualize PPopt result
}
\details{
Visualize the result of projection pursuit optimization
}
\examples{
data(iris)
PPoptViz(LDAopt(iris[,5],iris[,1:4],q=1))
PPoptViz(LDAopt(iris[,5],iris[,1:4],q=2))
}
\references{
Lee, EK(2017) 
PPtreeViz: R Package for Visualization of projection Pursuit Classification 
Tree, Journal of Statistical Software.
}
\keyword{tree}

