% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPoptViz.R
\name{PPoptViz}
\alias{PPoptViz}
\title{PPopt visualization}
\usage{
PPoptViz(PPoptOBJ)
}
\arguments{
\item{PPoptOBJ}{PPoptim object. result from LDAopt, PDAopt, and PPopt}
}
\description{
Visualize PPopt result
}
\details{
Visualize the result of projection pursuit optimization
}
\examples{
data(iris)
PPoptViz(LDAopt(iris[,5],iris[,1:4],q=1))
PPoptViz(LDAopt(iris[,5],iris[,1:4],q=2))
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005) 
Projection Pursuit for Exploratory Supervised Classification, 
Journal of Computational and Graphical Statistics, 14(4):831-846.
}
\keyword{tree}

