% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPtreeVizPlot.R
\name{PPtreeNode.Viz}
\alias{PPtreeNode.Viz}
\title{Finding PP tree structure using LDA index}
\usage{
PPtreeNode.Viz(PPtreeOBJ,node.id,Rule)
}
\arguments{
\item{PPtreeOBJ}{pptree object}

\item{node.id}{node ID}

\item{Rule}{cutoff rule}
}
\description{
Find tree structure using linear discriminant in each split.
}
\examples{
data(iris)
Tree.result <- LDA.Tree(iris[,5],iris[,1:4])
Tree.result
PPtreeNode.Viz(Tree.result,1,1)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013)
PPtree: Projection pursuit classification tree,
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

