% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printPPtree.R
\name{print.PPtree}
\alias{print.PPtree}
\title{Finding the 1D-optimal projection with PDA index}
\usage{
print.PPtree(PPtreeOBJ,coef.print=FALSE,cutoff.print=FALSE)
}
\arguments{
\item{PPtreeOBJ}{PPtree object}

\item{coef.print}{print projection coefficient in each node when TRUE}

\item{cutoff.print}{print cutoff values in each node when TRUE}
}
\description{
Find the 1D-optimal projection using PDA index for classification
}
\examples{
data(iris)
Tree.result <- LDA.Tree.viz(iris[,5],iris[,1:4])
Tree.result
print(Tree.result,coef.print=TRUE,cutoff.print=TRUE)
}
\references{
Lee, E.-K., Cook, D. (2009) A Projection Pursuit Index for Large p Small n Data,
Statistics and Computing,
\url{http://www.springerlink.com/content/g47n0n342761838m/#?p=d2ff5a7b69eb45ef8abf7ef3aba69557&pi=3}.
}
\keyword{tree}

